+++ packages/core/contracts/insured-bridge/BridgePool.sol
@@ -13,7 +13,6 @@ import "../common/implementation/AncillaryData.sol";
 import "../common/implementation/Testable.sol";
 import "../common/implementation/FixedPoint.sol";
 import "../common/implementation/Lockable.sol";
-import "../common/implementation/ExpandedERC20.sol";
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
@@ -156,11 +155,6 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
     event RelaySettled(bytes32 indexed depositHash, address indexed caller, RelayData relay);
     event BridgePoolAdminTransferred(address oldAdmin, address newAdmin);
 
-    modifier onlyFromOptimisticOracle() {
-        require(msg.sender == address(optimisticOracle), "Caller must be OptimisticOracle");
-        _;
-    }
-
     /**
      * @notice Construct the Bridge Pool.
      * @param _lpTokenName Name of the LP token to be deployed by this contract.
@@ -168,6 +162,7 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
      * @param _bridgeAdmin Admin contract deployed alongside on L1. Stores global variables and has owner control.
      * @param _l1Token Address of the L1 token that this bridgePool holds. This is the token LPs deposit and is bridged.
      * @param _lpFeeRatePerSecond Interest rate payment that scales the amount of pending fees per second paid to LPs.
+     * @param _isWethPool Toggles if this is the WETH pool. If it is then can accept ETH and wrap to WETH for the user.
      * @param _timer Timer used to synchronize contract time in testing. Set to 0x000... in production.
      */
     constructor(
@@ -196,7 +191,7 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
      *************************************************/
 
     /**
-     * @notice Add liquidity to the bridge pool. Pulls l1Token from the callers wallet. The caller is sent back a
+     * @notice Add liquidity to the bridge pool. Pulls l1Token from the caller's wallet. The caller is sent back a
      * commensurate number of LP tokens (minted to their address) at the prevailing exchange rate.
      * @dev The caller must approve this contract to transfer `l1TokenAmount` amount of l1Token if depositing ERC20.
      * @dev The caller can deposit ETH which is auto wrapped to WETH. This can only be done if: a) this is the Weth pool
@@ -222,7 +217,7 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
     }
 
     /**
-     * @notice Removes liquidity to the bridge pool. Burns lpTokenAmount LP tokens from the callers wallet. The caller
+     * @notice Removes liquidity from the bridge pool. Burns lpTokenAmount LP tokens from the caller's wallet. The caller
      * is sent back a commensurate number of l1Tokens at the prevailing exchange rate.
      * @dev The caller does not need to approve the spending of LP tokens as this method directly uses the burn logic.
      * @dev Reentrancy guard not added to this function because this indirectly calls sync() which is guarded.
@@ -258,9 +253,13 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
      * @dev Caller must have approved this contract to spend the total bond + amount - fees for `l1Token`.
      * @param depositData the deposit data struct containing all the user's deposit information.
      * @param realizedLpFeePct LP fee calculated off-chain considering the L1 pool liquidity at deposit time, before
-     *      quoteTimestamp. The OO acts to verify the correctness of this realized fee. Can not exceed 50%.
+     *      quoteTimestamp. The OO acts to verify the correctness of this realized fee. Cannot exceed 50%.
      */
     function relayAndSpeedUp(DepositData memory depositData, uint64 realizedLpFeePct) public nonReentrant() {
+        // If deposit quote time is before this contract's deployment, then clients will not be able to compute
+        // a realized LP fee %, so we should block such relays.
+        require(depositData.quoteTimestamp >= deploymentTimestamp, "Deposit before deployment");
+
         // If no pending relay for this deposit, then associate the caller's relay attempt with it.
         uint32 priceRequestTime = uint32(getCurrentTime());
 
@@ -269,7 +268,8 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
         require(
             depositData.slowRelayFeePct <= 0.25e18 &&
                 depositData.instantRelayFeePct <= 0.25e18 &&
-                realizedLpFeePct < 0.5e18
+                realizedLpFeePct <= 0.5e18,
+            "Invalid fees"
         );
 
         // Check if there is a pending relay for this deposit.
@@ -382,15 +382,20 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
      * @dev Caller must have approved this contract to spend the total bond + amount - fees for `l1Token`.
      * @param depositData the deposit data struct containing all the user's deposit information.
      * @param realizedLpFeePct LP fee calculated off-chain considering the L1 pool liquidity at deposit time, before
-     *      quoteTimestamp. The OO acts to verify the correctness of this realized fee. Can not exceed 50%.
+     *      quoteTimestamp. The OO acts to verify the correctness of this realized fee. Cannot exceed 50%.
      */
     function relayDeposit(DepositData memory depositData, uint64 realizedLpFeePct) public nonReentrant() {
+        // If deposit quote time is before this contract's deployment, then clients will not be able to compute
+        // a realized LP fee %, so we should block such relays.
+        require(depositData.quoteTimestamp >= deploymentTimestamp, "Deposit before deployment");
+
         // The realizedLPFeePct should never be greater than 0.5e18 and the slow and instant relay fees should never be
         // more than 0.25e18 each. Therefore, the sum of all fee types can never exceed 1e18 (or 100%).
         require(
             depositData.slowRelayFeePct <= 0.25e18 &&
                 depositData.instantRelayFeePct <= 0.25e18 &&
-                realizedLpFeePct < 0.5e18
+                realizedLpFeePct <= 0.5e18,
+            "Invalid fees"
         );
 
         // Check if there is a pending relay for this deposit.
@@ -525,7 +530,7 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
         // At this point there are two possible cases:
         // - This was a slow relay: In this case, a) pay the slow relayer their reward and b) pay the l1Recipient of the
         //      amount minus the realized LP fee and the slow Relay fee. The transfer was not sped up so no instant fee.
-        // - This was a instant relay: In this case, a) pay the slow relayer their reward and b) pay the instant relayer
+        // - This was an instant relay: In this case, a) pay the slow relayer their reward and b) pay the instant relayer
         //      the full bridging amount, minus the realized LP fee and minus the slow relay fee. When the instant
         //      relayer called speedUpRelay they were docked this same amount, minus the instant relayer fee. As a
         //      result, they are effectively paid what they spent when speeding up the relay + the instantRelayFee.
@@ -820,20 +825,7 @@ contract BridgePool is Testable, BridgePoolInterface, ERC20, Lockable {
     }
 
     function _getDepositHash(DepositData memory depositData) private view returns (bytes32) {
-        return
-            keccak256(
-                abi.encode(
-                    depositData.chainId,
-                    depositData.depositId,
-                    depositData.l1Recipient,
-                    depositData.l2Sender,
-                    address(l1Token),
-                    depositData.amount,
-                    depositData.slowRelayFeePct,
-                    depositData.instantRelayFeePct,
-                    depositData.quoteTimestamp
-                )
-            );
+        return keccak256(abi.encode(depositData, address(l1Token)));
     }
 
     // Proposes new price of True for relay event associated with `customAncillaryData` to optimistic oracle. If anyone