+++ packages/financial-templates-lib/src/price-feed/CreatePriceFeed.ts
@@ -34,7 +34,8 @@ import type { Logger } from "winston";
 import { NetworkerInterface } from "./Networker";
 import { PriceFeedInterface } from "./PriceFeedInterface";
 import { isDefined } from "../types";
-import { InsuredBridgeL1Client, InsuredBridgeL2Client, getL2DepositBoxAddress } from "..";
+import { InsuredBridgeL1Client, InsuredBridgeL2Client } from "..";
+import type { BlockTransactionBase } from "web3-eth";
 
 interface Block {
   number: number;
@@ -503,7 +504,7 @@ export async function createPriceFeed(
     const l2Client = new InsuredBridgeL2Client(
       logger,
       l2Web3,
-      await getL2DepositBoxAddress(providedWeb3, config.l2NetId, config.bridgeAdminAddress),
+      await l1Client.getL2DepositBoxAddress(config.l2NetId),
       config.l2NetId,
       currentL2Block - l2BlockLookback
     );
@@ -643,14 +644,14 @@ export async function createPriceFeed(
   }
 }
 
-const _blockFinderWorkaround = (web3: Web3) => BlockFinder((number: number | string) => web3.eth.getBlock(number));
-type BlockFinder = ReturnType<typeof _blockFinderWorkaround>;
-
 // Simple function to grab a singleton instance of the blockFinder to share the cache.
-const getSharedBlockFinder: { (web3: Web3): BlockFinder; blockFinder?: BlockFinder } = (web3: Web3): BlockFinder => {
+const getSharedBlockFinder: {
+  (web3: Web3): BlockFinder<BlockTransactionBase>;
+  blockFinder?: BlockFinder<BlockTransactionBase>;
+} = (web3: Web3): BlockFinder<BlockTransactionBase> => {
   // Attach the blockFinder to this function.
   if (!getSharedBlockFinder.blockFinder) {
-    getSharedBlockFinder.blockFinder = BlockFinder((number: number | string) => web3.eth.getBlock(number));
+    getSharedBlockFinder.blockFinder = new BlockFinder<BlockTransactionBase>(web3.eth.getBlock);
   }
   return getSharedBlockFinder.blockFinder;
 };