+++ packages/insured-bridge-relayer/test/Relayer.ts
@@ -121,7 +121,6 @@ describe("Relayer.ts", function () {
       .send({ from: l1Owner });
 
     l1Token = await ERC20.new("TESTERC20", "TESTERC20", 18).send({ from: l1Owner });
-    deployTimestamps = { [l1Token.options.address]: (await l1Timer.methods.getCurrentTime().call()).toString() };
     await l1Token.methods.addMember(TokenRolesEnum.MINTER, l1Owner).send({ from: l1Owner });
     await collateralWhitelist.methods.addToWhitelist(l1Token.options.address).send({ from: l1Owner });
     await store.methods.setFinalFee(l1Token.options.address, { rawValue: finalFee }).send({ from: l1Owner });
@@ -193,6 +192,8 @@ describe("Relayer.ts", function () {
       )
       .send({ from: l1Owner });
 
+    deployTimestamps = { [l1Token.options.address]: (await l1Timer.methods.getCurrentTime().call()).toString() };
+
     await bridgeDepositBox.methods
       .whitelistToken(l1Token.options.address, l2Token.options.address, bridgePool.options.address)
       .send({ from: l2BridgeAdminImpersonator });
@@ -240,7 +241,7 @@ describe("Relayer.ts", function () {
         amount: depositAmount,
         slowRelayFeePct: defaultSlowRelayFeePct,
         instantRelayFeePct: defaultInstantRelayFeePct,
-        quoteTimestamp: 1,
+        quoteTimestamp: Number((await bridgePool.methods.getCurrentTime().call()).toString()),
         depositContract: bridgeDepositBox.options.address,
       };
 
@@ -276,7 +277,7 @@ describe("Relayer.ts", function () {
     });
     it("Correctly decides when to slow relay", async function () {
       // The only time the relayer should decide to do a slow relay is when: a) the relayer has enough tokens, b) the
-      // deposit has had no other relayer pick it up and c) the deposit contains a instantRelayFeePct set to 0.
+      // deposit has had no other relayer pick it up and c) the deposit contains an instantRelayFeePct set to 0.
 
       // Mint tokens, set ClientRelayState to Any and update instantRelayFeePct.
       await l1Token.methods.mint(l1Relayer, toBN(defaultProposerBondPct).muln(2)).send({ from: l1Owner });
@@ -405,8 +406,8 @@ describe("Relayer.ts", function () {
 
       // Make a deposit on L2 and check the bot relays it accordingly.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const relaytime = await bridgePool.methods.getCurrentTime().call();
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const relayTime = await bridgePool.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       const depositData = {
         chainId,
         depositId: "0",
@@ -447,7 +448,7 @@ describe("Relayer.ts", function () {
       assert.isTrue(lastSpyLogIncludes(spy, "Slow Relay executed"));
 
       // Advance time such that relay has expired and check that bot correctly identifies it as expired.
-      const expirationTime = Number(relaytime.toString()) + defaultLiveness;
+      const expirationTime = Number(relayTime.toString()) + defaultLiveness;
       await bridgePool.methods.setCurrentTime(expirationTime).send({ from: l1Owner });
       await l1Client.update();
       assert.equal(
@@ -521,7 +522,7 @@ describe("Relayer.ts", function () {
     it("Can correctly detect and produce speedup relays", async function () {
       // Make a deposit on L2 and relay it. Then, check the relayer picks this up and speeds up the relay.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const currentBlockTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const currentBlockTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -574,7 +575,7 @@ describe("Relayer.ts", function () {
     it("Can correctly instantly relay deposits", async function () {
       // Make a deposit on L2 and see that the relayer correctly sends a slow relay and speedup it in the same tx.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const currentBlockTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const currentBlockTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -603,8 +604,8 @@ describe("Relayer.ts", function () {
       // Make a deposit on L2 and relay it with invalid relay params. The relayer should detect that the relay params
       // are invalid and skip it.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const relaytime = await bridgePool.methods.getCurrentTime().call();
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const relayTime = await bridgePool.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -647,7 +648,7 @@ describe("Relayer.ts", function () {
 
       // Advance time such that relay has expired and check that bot correctly identifies it as expired. Even if the
       // relay params are invalid, post-expiry its not disputable.
-      const expirationTime = Number(relaytime.toString()) + defaultLiveness;
+      const expirationTime = Number(relayTime.toString()) + defaultLiveness;
       await bridgePool.methods.setCurrentTime(expirationTime).send({ from: l1Owner });
       await Promise.all([l1Client.update()]);
       assert.equal((await Promise.allSettled(relayer.checkForPendingDepositsAndRelay())).length, 1);
@@ -673,29 +674,6 @@ describe("Relayer.ts", function () {
       await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Relayer });
       assert.equal((await Promise.allSettled(relayer.checkForPendingDepositsAndRelay())).length, 0);
       assert.isTrue(lastSpyLogIncludes(spy, "Deposit quote time < bridge pool deployment"));
-
-      // Relay the deposit from another slow relayer, and check that the bot skips any attempt to speed up the relay
-      // since it cannot verify its realized LP fee %.
-      await l1Token.methods.mint(l1Owner, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      await bridgePool.methods
-        .relayDeposit(
-          {
-            chainId: chainId,
-            depositId: "0",
-            l2Sender: l2Depositor,
-            l1Recipient: l2Depositor,
-            amount: depositAmount,
-            slowRelayFeePct: defaultSlowRelayFeePct,
-            instantRelayFeePct: defaultInstantRelayFeePct,
-            quoteTimestamp: quoteTime,
-          },
-          calculateRealizedLpFeePct(rateModel, toBNWei("0"), toBNWei("0.01")) // compute the expected fee for 1% utilization
-        )
-        .send({ from: l1Owner });
-      await Promise.all([l1Client.update(), l2Client.update()]);
-      assert.equal((await Promise.allSettled(relayer.checkForPendingDepositsAndRelay())).length, 0);
-      assert.isTrue(lastSpyLogIncludes(spy, "Deposit quote time < bridge pool deployment"));
     });
   });
   describe("Settle Relay transaction functionality", () => {
@@ -718,7 +696,7 @@ describe("Relayer.ts", function () {
 
       // Make a deposit on L2, relay it, advance time and check the bot settles it accordingly.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const currentBlockTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const currentBlockTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -834,7 +812,7 @@ describe("Relayer.ts", function () {
       // Make a deposit on L2 and relay it with invalid relay params. The disputer should detect that the relay params
       // are invalid and dispute it.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -941,7 +919,7 @@ describe("Relayer.ts", function () {
       // Make a deposit on L2 and relay it with invalid relay params. The disputer should detect that the relay params
       // are invalid and dispute it.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -992,7 +970,7 @@ describe("Relayer.ts", function () {
       // Relay a deposit that doesn't exist on-chain
       await l1Token.methods.mint(l1Owner, toBN(depositAmount).muln(2)).send({ from: l1Owner });
       await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgePool.methods
         .relayDeposit(
           {
@@ -1026,9 +1004,8 @@ describe("Relayer.ts", function () {
       assert.isTrue(lastSpyLogIncludes(spy, "No pending relays"));
     });
     it("Before disputing relays for deposits it cannot find, first tries to find deposit in new blocksearch", async function () {
-      // First send a valid deposit at the default quote time.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = Number((await web3.eth.getBlock("latest")).timestamp);
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -1060,10 +1037,11 @@ describe("Relayer.ts", function () {
         l1Relayer,
         whitelistedChainIds,
         deployTimestamps,
-        defaultLookbackWindow
+        1 // Use small lookback window to test that the back up block search loop runs at least a few times before
+        // finding the deposit.
       );
       await Promise.all([l1Client.update(), l2Client.update()]);
-      assert.equal((await Promise.allSettled(relayer.checkForPendingDepositsAndRelay())).length, 0);
+      await relayer.checkForPendingDepositsAndRelay();
       assert.isTrue(lastSpyLogIncludes(spy, "No relayable deposits"));
 
       // Now, relay the deposit.
@@ -1106,7 +1084,7 @@ describe("Relayer.ts", function () {
     it("Ignores relay for different whitelisted chain ID than the one set on L2 client", async function () {
       await l1Token.methods.mint(l1Owner, toBN(depositAmount).muln(2)).send({ from: l1Owner });
       await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
 
       // This relay should be disputed because the deposit doesn't exist on L2, but the disputer should skip it because
       // the L2 client is set for a different chain ID than the one included on the relay, and the chain ID
@@ -1140,7 +1118,7 @@ describe("Relayer.ts", function () {
     it("Does not dispute valid relay data that contains a valid deposit hash", async function () {
       // Make a deposit on L2 and relay it with valid relay params.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -1236,7 +1214,7 @@ describe("Relayer.ts", function () {
 
       // Make a deposit on L2 and relay it with valid relay params.
       await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      const quoteTime = await bridgeDepositBox.methods.getCurrentTime().call();
+      const quoteTime = (await web3.eth.getBlock("latest")).timestamp;
       await bridgeDepositBox.methods
         .deposit(
           l2Depositor,
@@ -1310,49 +1288,6 @@ describe("Relayer.ts", function () {
       await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Relayer });
       assert.equal((await Promise.allSettled(_relayer.checkForPendingRelaysAndDispute())).length, 1);
       assert.isTrue(lastSpyLogIncludes(spy, "Disputed pending relay"));
-    });
-    it("Disputes relays that bot cannot compute realized LP fee % for", async function () {
-      // Deposit using quote time prior to deploy timestamp for this L1 token.
-      const quoteTime = Number(deployTimestamps[l1Token.options.address]) - 1;
-      await l2Token.methods.approve(bridgeDepositBox.options.address, depositAmount).send({ from: l2Depositor });
-      await bridgeDepositBox.methods
-        .deposit(
-          l2Depositor,
-          l2Token.options.address,
-          depositAmount,
-          defaultSlowRelayFeePct,
-          defaultInstantRelayFeePct,
-          quoteTime.toString()
-        )
-        .send({ from: l2Depositor });
-
-      // Relay the deposit from another slow relayer, and check that the bot disputes the relay
-      // since its quote time < deploy timestamp for the pool.
-      await l1Token.methods.mint(l1Owner, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      await bridgePool.methods
-        .relayDeposit(
-          {
-            chainId: chainId,
-            depositId: "0",
-            l2Sender: l2Depositor,
-            l1Recipient: l2Depositor,
-            amount: depositAmount,
-            slowRelayFeePct: defaultSlowRelayFeePct,
-            instantRelayFeePct: defaultInstantRelayFeePct,
-            quoteTimestamp: quoteTime,
-          },
-          calculateRealizedLpFeePct(rateModel, toBNWei("0"), toBNWei("0.01")) // compute the expected fee for 1% utilization
-        )
-        .send({ from: l1Owner });
-      await Promise.all([l1Client.update(), l2Client.update()]);
-      await l1Token.methods.mint(l1Relayer, toBN(depositAmount).muln(2)).send({ from: l1Owner });
-      await l1Token.methods.approve(bridgePool.options.address, toBN(depositAmount).muln(2)).send({ from: l1Relayer });
-      assert.equal((await Promise.allSettled(relayer.checkForPendingRelaysAndDispute())).length, 1);
-      const targetLog = spy.getCalls().filter((_log: any) => {
-        return _log.lastArg.message.includes("Deposit quote time < bridge pool deployment");
-      });
-      assert.equal(targetLog.length, 1);
     });
   });
 });